fix(bq_driver): Enable masn, asan for internal dir unit test

+++ google/cloud/odbc/bq_driver/internal/odbc_internal_commons.cc
@@ -1178,7 +1178,7 @@ odbc_internal::StatusRecordOr<std::string> GetDataTypeInStr(BQDataType type) {
       return std::string("NULL");
     default:
       std::string err_msg = "Invalid BQ Data Type: ";
-      err_msg.append(std::to_string(type));
+      err_msg.append(std::to_string(static_cast<int>(type)));
       LOG(ERROR) << "GetDataTypeInStr:: " << err_msg;
       return StatusRecord{SQLStates::k_HY000(), err_msg};
   }
